requirements.txt
Django==4.2.8
gunicorn==20.1.0
psycopg2==2.9.9
djangorestframework==3.14.0